Move tools::IoError -> meta::error::IoError
